pyproject.toml
[tool.poetry]
name = "mlops-neechev"
version = "0.1.0"
description = ""
authors = ["Fakazuma <me@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25"
click = "^8.1.7"
timm = "^0.9.7"
torch = "2.0.0"
torchvision = "0.15.1"
fire = "^0.5.0"
dvc = "^3.27.0"
dvc-gdrive = "^2.20.0"
hydra-core = "1.3"
lightning = "^2.1.0"
transformers = "^4.34.1"
torchmetrics = "^1.2.0"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
tritonclient = "^2.41.0"
opencv-python = "^4.8.1.78"
gevent = "^23.9.1"
geventhttpclient = "^2.0.11"
google-api-python-client = "^2.111.0"

[tool.poetry.group.dev.dependencies]
setuptools = "^68.2.2"
pre-commit = "^3.4.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
